requirements.txt
tensorflow-gpu==2.3.0
tensorflow_datasets==3.2.1
cython==0.29.21
jupyterlab==2.2.2
notebook==6.1.0
#rasa_nlu==0.15.1
#sklearn_crfsuite==0.3.6
pydot==1.4.1
easydict==1.9
pandas==1.1.0